ikarus-backend/backend/requirements.txt
fastapi
uvicorn
pandas
numpy
sentence-transformers
faiss-cpu
pydantic